requirements.txt
funcsigs  # Python2 only
httplib2
m2crypto
m2ext
mock
munch
pdc-client
pyOpenSSL
python-fedora
# qpid-python  # Python2 only; not necessary; just to hide warning messages
six
sqlalchemy
futures  # Python 2 only
Flask
Flask-Migrate
Flask-SQLAlchemy
Flask-Script
requests
jinja2
pyldap
